Please squash this commit in the pull request

+++ drupal/scripts/deploy.sh
@@ -57,6 +57,8 @@ fi
 # Copy all items from /drupal-modules-contrib and /drupal-themes-contrib to
 # modules/contrib and themes/contrib.
 # See comments in ./docker-resources/drupal/build-drupal.sh for more details.
-cp -r /drupal-modules-contrib/* modules/contrib/
-cp -r /drupal-themes-contrib/* themes/contrib/
-rm -rf /drupal-modules-contrib /drupal-themes-contrib
+if [ -d /drupal-modules-contrib ]; then
+  cp -r /drupal-modules-contrib/* modules/contrib/
+  cp -r /drupal-themes-contrib/* themes/contrib/
+  rm -rf /drupal-modules-contrib /drupal-themes-contrib
+fi
